changed exceptional package to pojo added one to one hibernate link between group and teacher
