+++ model/src/main/java/by/pojo/Admin.java
@@ -1,4 +1,4 @@
-package exceptions.pojo;
+package by.pojo;
 
 import lombok.*;
