+++ model/src/main/java/by/pojo/Group.java
@@ -1,9 +1,7 @@
-package exceptions.pojo;
+package by.pojo;
 
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
-import com.fasterxml.jackson.annotation.ObjectIdGenerator;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import lombok.*;
