+++ model/src/main/java/by/pojo/Student.java
@@ -1,43 +1,45 @@
-package exceptions.pojo;
+package by.pojo;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import lombok.*;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
-@Data
-@With
 @AllArgsConstructor
 @NoArgsConstructor
+@Data
+@With
 @JsonIdentityInfo(generator = ObjectIdGenerators.IntSequenceGenerator.class, property = "id")
-public class Teacher extends AbstractEntity {
-    private List<Salary> salaries = new ArrayList<>();
+public class Student extends AbstractEntity {
     private String login;
     private String password;
     private String first_name;
     private String last_name;
-    private Group group;
+    private Set<Group> groups = new HashSet<>();
     private Integer age;
 
-    public Teacher(int id, String login, String password, String first_name, String last_name, int age) {
+    public Student(int id, String login, String password, String first_name, String last_name, int age, Set<Group> groups) {
         super(id);
         this.login = login;
         this.password = password;
         this.first_name = first_name;
         this.last_name = last_name;
         this.age = age;
+        this.groups = groups;
     }
 
     @Override
-    public Teacher withId(Integer id) {
+    public Student withId(Integer id) {
         setId(id);
+        return this;
+    }
+
+    public Student withGroup(Group group) {
+        groups.add(group);
         return this;
     }
 }
