+++ model/src/main/java/by/pojo/Theme.java
@@ -1,4 +1,4 @@
-package exceptions.pojo;
+package by.pojo;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
