+++ pom.xml
@@ -60,12 +60,11 @@
                 <artifactId>jackson-databind</artifactId>
                 <version>2.12.0</version>
             </dependency>
-<!--            <dependency>-->
-<!--                <groupId>com.fasterxml.jackson.core</groupId>-->
-<!--                <artifactId>jackson-annotations</artifactId>-->
-<!--                <version>2.12.0</version>-->
-<!--                <scope>provided</scope>-->
-<!--            </dependency>-->
+            <dependency>
+                <groupId>org.hibernate</groupId>
+                <artifactId>hibernate-core</artifactId>
+                <version>5.4.28.Final</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
