+++ repository/src/main/java/all/repos/AbstractRepositoryPostgres.java
@@ -1,8 +1,7 @@
 package all.repos;
 
 import exceptions.DatabaseException;
-import exceptions.pojo.AbstractEntity;
-import exceptions.pojo.Student;
+import by.pojo.AbstractEntity;
 import lombok.extern.slf4j.Slf4j;
 
 import java.sql.Connection;
