+++ repository/src/main/java/all/repos/GroupRepository.java
@@ -1,6 +1,6 @@
 package all.repos;
 
-import exceptions.pojo.Group;
+import by.pojo.Group;
 
 public interface GroupRepository extends Repository<Group> {
 }
