+++ repository/src/main/java/all/repos/TeacherRepository.java
@@ -1,6 +1,6 @@
 package all.repos;
 
-import exceptions.pojo.Teacher;
+import by.pojo.Teacher;
 
 public interface TeacherRepository extends Repository<Teacher>{
 }
