+++ repository/src/main/java/all/repos/TeacherRepositoryPostgres.java
@@ -1,8 +1,8 @@
 package all.repos;
 
-import exceptions.pojo.Group;
-import exceptions.pojo.Student;
-import exceptions.pojo.Teacher;
+import by.pojo.Group;
+import by.pojo.Student;
+import by.pojo.Teacher;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
