+++ service/src/main/java/exceptions/service/TeacherServiceImpl.java
@@ -4,8 +4,8 @@ import all.repos.GroupRepository;
 import all.repos.RepositoryFactory;
 import all.repos.TeacherRepository;
 import all.repos.TeacherRepositoryPostgres;
-import exceptions.pojo.Group;
-import exceptions.pojo.Teacher;
+import by.pojo.Group;
+import by.pojo.Teacher;
 
 import java.util.Optional;
 import java.util.Set;
